+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [v1.11.3 - 2024-02-16]
 ### Changed
 - dblib: use bitnami/openldap as test container
+- maillib: use mailserver:13.2.0 and refactor tests to fit there
 
 ## [v1.11.2 - 2024-02-13]
 ### New